requirements.txt
flask
pillow
pytesseract
ibm-watson
